chnage_languages(Ver2.0)
Added Language buttons
Added design
