+++ package.json
@@ -7,9 +7,12 @@
     "@testing-library/react": "^11.2.5",
     "@testing-library/user-event": "^12.8.3",
     "ga-4-react": "^0.1.281",
+    "i18next": "^20.3.1",
     "react": "^17.0.1",
     "react-dom": "^17.0.1",
-    "react-scripts": "^4.0.3",
+    "react-i18next": "^11.10.0",
+    "react-scripts": "^1.1.5",
+    "ress": "^4.0.0",
     "web-vitals": "^1.1.0"
   },
   "scripts": {
@@ -45,7 +48,7 @@
     "eslint-plugin-import": "^2.23.2",
     "eslint-plugin-node": "^11.1.0",
     "eslint-plugin-promise": "^4.3.1",
-    "gh-pages": "^3.1.0",
+    "gh-pages": "^1.1.0",
     "standard": "^16.0.3"
   }
 }
