+++ src/locales/en.json
@@ -1,0 +1,23 @@
+{
+    "hello": "hello",
+    "最難テトリス ”ナカヤマン” Ver2.0": "The Hardest Tetris ”NAKAYAMAN”Ver2.0",
+    "スコア5000点以上とって、ナカヤマンをぶっ倒せ！！！": "Defeat 'NAKAYAMAN' with a score of 5000 points higher !!!",
+    "次のブロック": "Next Block",
+
+    "スコア" : "Score",
+    "レベル" : "Level",
+    "ナカヤマン" : "NAKAYAMAN",
+    "勝利判定" : "Winner Judgement",
+
+    "5000点以上とるとわ。。。 いっぱい遊んで、くれたんやな、、、ありがとう！！！完敗や" : "You got more than 5000 points. .. .. Thank you for playing a lot! !! !! Complete defeat",
+    "あなたの勝ち" : "You are winner",
+    "がははは、100年早いわ！出直してきんしゃい！" : "Gahaha, 100 years early! Come back again!",
+    "あなたの負け" : "You are Loser",
+    "さて、お手並み拝見させて貰おうか" : "Well...Let me see how you are",
+    "ほほう！やり方は、知っているみたいだな。それじゃ行くぞ" : "O oh! You seem to know how to play it. Well here we go",
+    "むむむ、なかなかうまいな" : "Hmmm, pretty good",
+    "ま、ま、まだ大丈夫だ" : "i...i...it doesn't matter",
+    "こ、このままだと、負けてしまう。。あれを準備しよう。。。" : "Ohhh...If you keep this, I will lose. .. I have to prepare that...",
+    "奥義発動" : "Mystery activation",
+    "勝たせない！" : "I never make you win!"
+  }
