+++ src/locales/ja.json
@@ -1,0 +1,23 @@
+{
+    "hello": "こんにちは",
+    "最難テトリス ”ナカヤマン” Ver2.0": "最難テトリス ”ナカヤマン Ver2.0”",
+    "スコア5000点以上とって、ナカヤマンをぶっ倒せ！！！": "スコア5000点以上とって、ナカヤマンをぶっ倒せ！！！",
+    "次のブロック": "次のブロック",
+    
+    "スコア" : "スコア",
+    "レベル" : "レベル",
+    "ナカヤマン" : "ナカヤマン",
+    "勝利判定" : "勝利判定",
+
+    "5000点以上とるとわ。。。 いっぱい遊んで、くれたんやな、、、ありがとう！！！完敗や" : "5000点以上とるとわ。。。 いっぱい遊んで、くれたんやな、、、ありがとう！！！完敗や",
+    "あなたの勝ち" : "あなたの勝ち",
+    "がははは、100年早いわ！出直してきんしゃい！" : "がははは、100年早いわ！出直してきんしゃい！",
+    "あなたの負け" : "あなたの負け",
+    "さて、お手並み拝見させて貰おうか" : "さて、お手並み拝見させて貰おうか",
+    "ほほう！やり方は、知っているみたいだな。それじゃ行くぞ" : "ほほう！やり方は、知っているみたいだな。それじゃ行くぞ",
+    "むむむ、なかなかうまいな" : "むむむ、なかなかうまいな",
+    "ま、ま、まだ大丈夫だ" : "ま、ま、まだ大丈夫だ",
+    "こ、このままだと、負けてしまう。。あれを準備しよう。。。" : "こ、このままだと、負けてしまう。。あれを準備しよう。。。",
+    "奥義発動" : "奥義発動",
+    "勝たせない！" : "勝たせない！"
+  }
